Move YAMLOpInfoCreator to the engine
