+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/yaml/impl/JavaFieldYAMLOpInfoCreator.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.engine.yaml;
+package org.scijava.ops.engine.yaml.impl;
 
 import java.lang.reflect.Field;
 import java.net.URI;
@@ -8,8 +8,9 @@ import java.util.Map;
 import org.scijava.common3.Classes;
 import org.scijava.ops.api.Hints;
 import org.scijava.ops.api.OpInfo;
-import org.scijava.ops.api.features.YAMLOpInfoCreator;
 import org.scijava.ops.engine.matcher.impl.OpFieldInfo;
+import org.scijava.ops.engine.yaml.AbstractYAMLOpInfoCreator;
+import org.scijava.ops.engine.yaml.YAMLOpInfoCreator;
 
 /**
  * A {@link YAMLOpInfoCreator} specialized for Java {@link Field}s.
@@ -24,8 +25,7 @@ public class JavaFieldYAMLOpInfoCreator extends AbstractYAMLOpInfoCreator {
 	}
 
 	@Override
-	OpInfo create(String identifier, String[] names, double priority,
-                  String version, Map<String, Object> yaml) throws Exception
+	protected OpInfo create(String identifier, String[] names, double priority, String version, Map<String, Object> yaml) throws Exception
 	{
 		// parse class
 		int clsIndex = identifier.indexOf('$');
