+++ scijava/scijava-ops-engine/src/main/resources/META-INF/services/org.scijava.discovery.Discoverer
@@ -1,3 +1,3 @@
 # -- DO NOT EDIT -- This file is autogenerated
 # Instead, modify and re-run <basedir>/bin/generate-meta-inf.sh from the top-level
-org.scijava.ops.engine.yaml.YAMLOpInfoDiscoverer
+org.scijava.ops.engine.yaml.impl.YAMLOpInfoDiscoverer
