--- scijava/scijava-ops-engine/src/main/resources/META-INF/services/org.scijava.ops.api.features.YAMLOpInfoCreator
@@ -1,5 +1,0 @@
-# -- DO NOT EDIT -- This file is autogenerated
-# Instead, modify and re-run <basedir>/bin/generate-meta-inf.sh from the top-level
-org.scijava.ops.engine.yaml.JavaClassYAMLOpInfoCreator
-org.scijava.ops.engine.yaml.JavaFieldYAMLOpInfoCreator
-org.scijava.ops.engine.yaml.JavaMethodYAMLInfoCreator
