+++ scijava/scijava-ops-engine/src/main/resources/META-INF/services/org.scijava.ops.engine.yaml.YAMLOpInfoCreator
@@ -1,0 +1,5 @@
+# -- DO NOT EDIT -- This file is autogenerated
+# Instead, modify and re-run <basedir>/bin/generate-meta-inf.sh from the top-level
+org.scijava.ops.engine.yaml.impl.JavaClassYAMLOpInfoCreator
+org.scijava.ops.engine.yaml.impl.JavaFieldYAMLOpInfoCreator
+org.scijava.ops.engine.yaml.impl.JavaMethodYAMLInfoCreator
