+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/yaml/impl/ops/YAMLClassOp.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.engine.yaml.ops;
+package org.scijava.ops.engine.yaml.impl.ops;
 
 import java.util.function.BiFunction;
